Refactor spectrum and VU level handling to use caller-allocated buffers
- Updated `cdsp_spectrum_t` to require caller-allocated frequency and magnitude arrays.
- Modified `cdsp_get_spectrum` to accept caller-allocated buffers and removed dynamic memory allocation within the function.
- Refactored VU level handling in the websocket server to allocate buffers in the caller context, improving memory management.
- Updated related tests to validate the new caller-allocated buffer approach for VU levels and spectrum analysis.
- Removed unnecessary free calls for dynamically allocated VU levels and spectrum data.

+++ Audio/spectrum_analyzer.h
@@ -86,8 +86,8 @@ void spectrum_analyzer_free(spectrum_analyzer_t* analyzer);
  */
 spectrum_status_t spectrum_analyzer_compute(spectrum_analyzer_t* analyzer,
                                             audio_history_buffer_t* buffer,
-                                            int channel, double min_freq,
-                                            double max_freq, size_t n_bins,
+                                            int channel, float min_freq,
+                                            float max_freq, size_t n_bins,
                                             size_t samplerate,
                                             spectrum_result_t* out_result);
